Move ui code into pookie/ui
