+++ src/main/java/pookie/Main.java
@@ -7,6 +7,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import pookie.ui.MainWindow;
 
 /**
  * A GUI for Duke using FXML.
