+++ src/main/java/pookie/ui/MainWindow.java
@@ -1,4 +1,4 @@
-package pookie;
+package pookie.ui;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -7,6 +7,7 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
+import pookie.Pookie;
 
 /**
  * Controller for the main GUI.
